Clean up some of the god awful ai generated volume code

+++ expo/modules/snow-audio-controls/android/src/main/java/com/simplepathstudios/snowgroove/audiocontrols/SnowAudioControlsModule.kt
@@ -114,7 +114,7 @@ class SnowAudioControlsModule : Module() {
             }
 
             Function("setVolume") { volume: Double ->
-                playbackService?.setVolumeLevel(volume.toFloat())
+                playbackService?.setVolumeLevel(volume)
             }
 
             Function("moveCurrentIndex") { amount: Int ->
